% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createIGD}
\alias{createIGD}
\title{Function to create an IGD database from a folder of .bed or .bed.gz files,
or a list of such folders}
\usage{
createIGD(iPath, oPath, igdName, binsize = 16384)
}
\arguments{
\item{iPath}{folder where your input files are stored}

\item{oPath}{the folder that the created IGD database will be stored}

\item{igdName}{the name you give to the IGD database (.igd will be added to it)}

\item{binsize}{the size in basepairs for the bin (block) used in the database:
usually 8192, 16384, 32768, ... as a power of 2; default 16384}
}
\value{
an igd database will be created in the specified folder
}
\description{
Function to create an IGD database from a folder of .bed or .bed.gz files,
or a list of such folders
}
\examples{
\dontrun{
library("IGDr")
iPath <- system.file("extdata", "rme3", package = "IGDr")
IGDr::createIGD(iPath, "testigd", "roadmap_b14")
}
}
