% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{createIGD_f}
\alias{createIGD_f}
\title{Function to create an IGD database from a list of source files (.bed or .bed.gz)}
\usage{
createIGD_f(iPath, oPath, igdName, binsize = 16384)
}
\arguments{
\item{iPath}{path to a txt file that lists the paths of all the source files}

\item{oPath}{the folder that the created IGD database will be stored}

\item{igdName}{the name you give to the IGD database (.igd will be added to it)}

\item{binsize}{the size in basepairs for the bin (block) used in the database:
usually 8192, 16384, 32768, ... as a power of 2}
}
\value{
an igd database will be created in the specified folder
}
\description{
Function to create an IGD database from a list of source files (.bed or .bed.gz)
}
