% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{get_binData}
\alias{get_binData}
\title{Function to get region data in the given bin of a given contig}
\usage{
get_binData(igdr, ichr, binID)
}
\arguments{
\item{igdr}{an igd database object (loaded)}

\item{ichr}{contig id}

\item{binID}{the bin number}
}
\value{
binData: vector of regions start, end, source_id
}
\description{
Function to get region data in the given bin of a given contig
}
\examples{
\dontrun{
binData <- get_binLen(igdr, 10, 123)}
}
