% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{get_binLen}
\alias{get_binLen}
\title{Function to get the number of regions in a given bin and a given contig}
\usage{
get_binLen(igdr, ichr, binID)
}
\arguments{
\item{igdr}{an igd database object (loaded)}

\item{ichr}{contig id}

\item{binID}{the bin number}
}
\value{
binLen
}
\description{
Function to get the number of regions in a given bin and a given contig
}
\examples{
binLen <- get_binLen(igdr, 10, 123)
}
