% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{get_binSize}
\alias{get_binSize}
\title{Function to get the bin size of an igd database}
\usage{
get_binSize(igdr)
}
\arguments{
\item{igdr}{an igd database object (loaded)}
}
\value{
binSize
}
\description{
Function to get the bin size of an igd database
}
\examples{
binSize <- get_binSize(igdr)
}
