% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{get_nCtgs}
\alias{get_nCtgs}
\title{Function to get the number of contigs in an igd database}
\usage{
get_nCtgs(igdr)
}
\arguments{
\item{igdr}{an igd database object (loaded)}
}
\value{
nCtgs: number of contigs
}
\description{
Function to get the number of contigs in an igd database
}
\examples{
\dontrun{
nCtgs <- get_ctgId(igdr)}
}
