% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{get_nFiles}
\alias{get_nFiles}
\title{Function to get the number of source files in an igd database}
\usage{
get_nFiles(igdr)
}
\arguments{
\item{igdr}{an igd database object (loaded)}
}
\value{
nCtgs: number of source files
}
\description{
Function to get the number of source files in an igd database
}
\examples{
\dontrun{
nFiles <- get_nFiles(igdr)}
}
