% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{IGDr}
\alias{IGDr}
\title{Function to open/load an igd database for search}
\usage{
IGDr(igd_file)
}
\arguments{
\item{igd_file}{the path to the igd database file}
}
\value{
an IGDr object
}
\description{
Function to open/load an igd database for search
}
\examples{
\dontrun{library(IGDr)
igd_file <- system.file("extdata", "rme3_igd/roadmap_b14.igd")
igdr <- IGDr(igd_file)}
}
