% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{search_1r}
\alias{search_1r}
\title{Function to search the igd database for a single query}
\usage{
search_1r(igdr, chrm, qs, qe)
}
\arguments{
\item{igdr}{an igd database object (loaded)}

\item{chrm}{the chromosome name of the query in format of chr1, chrX, chrY, chrM, ...}

\item{qs}{the start location of the query}

\item{qe}{the end location of the query}
}
\value{
hits: number of intersections to each database source file
}
\description{
Function to search the igd database for a single query
}
\examples{
\dontrun{
hits <- search_1r(igdr, "chr1", 1000000, 1100000)}
}
