% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{search_nr}
\alias{search_nr}
\title{Function to search the igd database for multiple queries}
\usage{
search_nr(igdr, n, chrm, qs, qe)
}
\arguments{
\item{igdr}{an igd database object (loaded)}

\item{n}{number of queries to be searched}

\item{chrm}{vector of chromosome names}

\item{qs}{vector of the start locations of the queries}

\item{qe}{vector of the end locations of the queries}
}
\value{
hits: number of intersections to each database source file
}
\description{
Function to search the igd database for multiple queries
}
\examples{
hits <- search_nr(igdr, chrm, qs, qe)
}
