% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IGDr.R
\name{search_qfile}
\alias{search_qfile}
\title{Function to search the igd database for a query set from a file}
\usage{
search_qfile(igdr, qfile)
}
\arguments{
\item{igdr}{an igd database object (loaded)}

\item{qfile}{path to the query file (file type of .bed or .bed.gz)}
}
\value{
hits: number of intersections to each database source file
}
\description{
Function to search the igd database for a query set from a file
}
\examples{
\dontrun{
hits <- search_qfile(igdr, "data/r10000.bed")}
}
